[c-run-time] MMAllocMisc/MMAllocFree uses unsigned long to match generated code.

+++ sources/lib/run-time/run-time.h
@@ -1173,8 +1173,8 @@ extern void primitive_replaceX
  * be freed explicitly.
  */
 
-extern void *MMAllocMisc(size_t size);
-extern void MMFreeMisc(void *p, size_t size);
+extern void *MMAllocMisc(unsigned long size);
+extern void MMFreeMisc(void *p, unsigned long size);
 extern void *dylan__malloc__misc(size_t size);
 extern void *dylan__malloc__ambig(size_t size);
 extern void *mps__malloc(size_t size);
